// src/sharc/types/Theme.ts
import * as Types from "./Sprites";

export type DefaultsType = {
    Line?: Types.LineProperties;
    NullSprite?: Types.NullSpriteProperties;
    Rect?: Types.RectProperties;
    Ellipse?: Types.EllipseProperties;
    BezierCurve?: Types.BezierCurveProperties;
    Path?: Types.PathProperties;
    Polygon?: Types.PolygonProperties;
    Star?: Types.StarProperties;
    TextSprite?: Types.TextProperties;
    ImageSprite?: Types.ImageProperties;
    LabelSprite?: Types.LabelProperties;
    ManagerSprite?: Types.ManagerProperties;
    FactorySprite?: Types.FactoryProperties;
    PolarWrapper?: Types.PolarWrapperProperties;
};